Implemented VFF Rebalancing.
Note that executing with 200+ nodes may bring you in an infinite loop.
A termination condition when no solution can be found should be implemented.
See line 286 for its cause. Without that line, the rebalancing falls into the same coloring.
That means the input coloring becomes the output coloring.

Main is set up for implementing the methods that enable writing down the coloring.